refactor(capabilities): split lifecycle into a submodule
Replace the 1695-line `lifecycle.rs` monolith with a `lifecycle/`
directory submodule split along its cohesion seams (ADR-0121), matching
the `tcp/` and `engine/` layout:

- `graph.rs` — the data-graph builder (`LifecycleGraphData` /
  `LifecycleGraphBuilder` type-states / `BuildError` / `finalize`) and
  its builder tests.
- `subscribers.rs` — the always-on `LifecycleMailboxExt` sender facade
  (both transports) and the native `broadcast_to_subscribers` fan-out.
- `settlement.rs` — the settlement-gated advance state machine
  (`PendingAdvance`, `Step`, `resolve_edge`, the EWMA latency tracking
  and force-complete fallback) and its edge/latency unit tests.
- `mod.rs` — the cap shell: module docs, re-exports preserving every
  public path, and the macro-indivisible `#[bridge(singleton)] mod
  native` with `LifecycleConfig`, the `LifecycleCapability` struct
  (fields relaxed to `pub(crate)`), `init`, the whole `#[actor] impl`,
  the read accessors, and the shared test fixtures.

Split-only: no kinds move and no `kinds.rs` is added — the substrate
scheduler dispatches `Tick` from the core, so the lifecycle kind family
stays in `aether-kinds` (the cycle rule, ADR-0121 §3). Behavior, the
public `crate::lifecycle::*` paths, and every kind id are unchanged.
